+++ bin/refresh-ctx-core-package.json.js
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 require = require('esm')(module)
-require('../src').refresh_ctx_core_package().then()
+require('../dist').refresh_ctx_core_package().then()