+++ bin/run-parallel--workspaces.js
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 require = require('esm')(module)
-require('../src').run_parallel_workspaces_cli().then()
+require('../dist').run_parallel_workspaces_cli().then()