+++ bin/sort-packages-exec.js
@@ -1,7 +1,7 @@
 #!/bin/env node
 require = require('esm')(module);
-const { exec, sort_packages_call } = require('../src')
-const cmd = process.argv.slice(2).join(' ')
+const { exec, sort_packages_call } = require('../dist')
+const cmd = process.argv.slice(process.argv.indexOf('--') + 1).join(' ')
 sort_packages_call(async (pkg)=>{
   try {
     const res = await exec(`cd ${pkg.path} && ${cmd}`)