+++ src/monorepo_thread_opts_T.d.ts
@@ -1,4 +1,0 @@
-export interface monorepo_thread_opts_T {
-    threads?: number;
-    package_name?: string | string[];
-}