+++ src/projects_.d.ts
@@ -1,2 +1,0 @@
-import type { project_T } from './project_T';
-export declare function projects_(): Promise<project_T[]>;