+++ src/run_parallel_workspaces.d.ts
@@ -1,2 +1,0 @@
-import type { monorepo_thread_opts_T } from './monorepo_thread_opts_T';
-export declare function run_parallel_workspaces(cmd_a: string[], opts?: monorepo_thread_opts_T): Promise<Record<string, string>>;