+++ src/stdout_h_package_name_.d.ts
@@ -1,1 +1,0 @@
-export declare function stdout_h_package_name_(package_name_a: string[], stdout_a2: string[]): Record<string, string>;