+++ src/writeFile.d.ts
@@ -1,3 +1,0 @@
-/// <reference types="node" />
-import fs from 'fs';
-export declare const writeFile: typeof fs.writeFile.__promisify__;